services/rules_engine/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pyyaml==6.0.1
markdown==3.5.1
python-multipart==0.0.6
google-generativeai==0.3.2
pypdf2==3.0.1
ebooklib==0.18
chromadb==0.4.18

